+++ src/db/IFloodRepository.h
@@ -21,10 +21,10 @@ public:
 
     virtual std::map<std::string, std::string> getAllStations() = 0;
     virtual bool stationExists(std::string stationName) = 0;
-    virtual std::vector<RiverReading> getRiverReadings(std::string startDate = "", uint16_t page = 1,
+    virtual std::vector<RiverReading> getRiverReadings(std::string startDate = "", int page = 1,
                                                        uint8_t pageSize = 12) const = 0;
     virtual std::vector<RainfallReading> getStationRainfallReadings(std::string stationName, std::string startDate = "",
-                                                                    uint16_t page = 1, uint8_t pageSize = 12) const = 0;
+                                                                    int page = 1, uint8_t pageSize = 12) const = 0;
   };
 } // namespace jbriggs::flood::db
